repactor: index.ts 파일을 추가하여 구조 개선
