+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header.tsx"
-import Footer from "./widgets/ui/Footer.tsx"
+import { Header, Footer } from "./widgets"
 import PostsManagerPage from "./pages/PostsManagerPage.tsx"
 
 const App = () => {
