+++ src/widgets/index.ts
@@ -1,0 +1,2 @@
+export { default as Header } from "./Header"
+export { default as Footer } from "./Footer"
